# Improve autovectorization of to_lowercase / to_uppercase functions