@jhorstmann 2024-04-11
r? @the8472 

The assembly for x86 and aarch64 can also be seen at https://rust.godbolt.org/z/x6T65nE8E