@Marcondiro 2024-05-05
Hi @jhorstmann do you have any benchmark for this? Thx!