@jhorstmann 2024-05-06
@Marcondiro only the microbenchmark included in this PR. On my machine (Intel i9-11900KB) the performance increases by nearly 3x. This is without any target-specific compiler flags, rerunning them now with:

```
./x.py bench library/alloc/ --stage 0 --test-args to_lowercase
```
Before
```
benchmarks:
    string::bench_to_lowercase 57.00ns/iter +/- 1.00ns
```
After
```
benchmarks:
    string::bench_to_lowercase 20.00ns/iter +/- 1.00ns
```