@Marcondiro 2024-05-07
Cool!
I tried on an M1 and the results are not as good
Before:
```
string::bench_to_lowercase 102.00ns/iter +/- 1.00ns
```
After:
```
string::bench_to_lowercase 99.00ns/iter +/- 1.00ns
```
Maybe autovec is not as powerful here idk