@jhorstmann 2024-05-08
Thanks for running the benchmarks, glad that there is no regression on arm. The improvement on x86 mostly comes from the usage of the `pmovmskb` instruction, the equivalent on arm requires multiple instructions with higher latency/lower throughput.