@Marcondiro 2024-05-08
Sorry, I was testing your solution against my fork that I am working on instead of upstream 🙈
The actual result is
Before:
```
string::bench_to_lowercase 45.00ns/iter +/- 0.00ns
```
After:
```
string::bench_to_lowercase 99.00ns/iter +/- 1.00ns
```