@jhorstmann 2024-05-09
Thanks again, I was able to reproduce this on an aws c7g instance. It seems there were some bounds check remaining in the generated assembly, which were removed by the compiler on x86_64 and also in the simplified version I checked in compiler explorer. Adding some `assume` fixes that and the benchmarks on c7g / Neoverse-V1 are now

Before
```
benchmarks:
    string::bench_to_lowercase 61.00ns/iter +/- 0.00ns
```
After
```
benchmarks:
    string::bench_to_lowercase 28.00ns/iter +/- 0.00ns
```